By changing outputclass to TanhLayer (anything but SoftMaxLayer), the training now works!
Refactored a large portion of the code: Almost everything that has to do with reading data from file has been moved to the data_moving file.
Additionally, a few methods (and Point, and ROI) has been moved to common, because they are used in the reading of data as well. This can be fixed by having classes as input, but...