Fixed bug in Wavelengths
Made parameters.py where all the info in wavelengths for the different
data sets reside.
Common, has gotten some new methods (indexing).
Wave changed the way wavelength info is added to the database; it is
done after all the points has been added. This is to fix the bug where
NEW wavelength info was added FOR EACH point! (making the wavelength
table very big), and very wrong.
 settings.py will contain settings for everything that has settings.
(the DATABASE is defined in there, by reading it from a separate file,
which is not included due to security reasons.)