Bug fix: add only wavelength info to the points of the given dataset.
Fixed a bug that would cause EVERY SINGLE point to change wavelength
information as more was added: the method in connector.py did not take
to which dataset the points belonged. This should have been fixed now.
Added docstrings to some of the methods.
Iteration over all the points in add_wavelength_to_points has been
changed so that it does not get the entire dataset at once. This might
be slower... Will check it later.

Also, started implementing methods for removing stuff, so that regions
of interest can be changed later.